tools/seo-expert/requirements.txt
# CoreLTB SEO Expert - Dependencies

# AI APIs
google-genai>=0.3.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0

# Embeddings & ML
numpy>=1.24.0
scikit-learn>=1.3.0

# CLI & Utils
click>=8.1.0
colorama>=0.4.6
rich>=13.7.0

# Web Dashboard
streamlit>=1.40.0

# Data Processing
pydantic>=2.5.0
python-dotenv>=1.0.0
